src/flash_cve_detector/requirements.txt
yara-python==3.5
psutil==5.0